Add files via upload
I am a seasoned Full Stack Web Developer with a diverse educational background that includes a Diploma in Database Administration, a Bachelor of Science in Natural Resources Management, and a Master of Science in Natural Resources Management with a specialization in Watershed Management. Currently, I am on an exciting academic journey as a Ph.D. candidate in Environment and Natural Resources Management.